:art: making the api nicer
