+++ packages/enzyme-utilities/src/index.ts
@@ -1,3 +1,2 @@
-export {default as wrappers} from './wrappers';
-export {default as createMount} from './mount';
+export * from './mount';
 export * from './trigger';
